refactor: 선 차트 페이지와 분리
